Switch back to Docker (#1)
Moving to K8s has been a good learning experience but it is overkill and has too many downsides:
- My single node has intermittent network issues for some reason, that causes pods to temporarily lose connections with each other (this causes problems with, for example, Authelia to Authelia DB communication). Could not fix despite trying for hours.
- I can't run Mullvad VPN on the node (somehow interferes with K8s networking (tested with Flannel and Calico). Could not fix despite trying for days.
- For some reason K8s causes a process called `gvfs-udisks2-volume-monitor` to run with high CPU usage 24/7 - not bad enough to be a showstopper but bad enough to be annoying and increase fan noise. Could not fix despite trying for hours.
- SELinux causes problems with volume mounts and Pod permissions. Took a long time to figure out what options to use to get around them without disabling it entirely.

Basic takeaway is that using a cluster administered by others is great once you learn how. Maintaining a cluster though, is a clusterfuck. Not worth it for a single node with weak hardware that doubles as a media center (thus using Fedora - an unusual OS for K8s).

This PR switches back to a much simpler Docker compose based system, with improvements on the old compose system based on learnings from the K8s experience. No more hardcoded secrets, for example, thus making it possible to store this on GitHub.